rebase master and improve code
